server/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
setuptools>=65.0.0
numpy>=1.21,<2.0
pandas==2.1.4
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
Pillow==10.1.0
openpyxl==3.1.2
